makesymlinks.sh Textrückmeldungen bearbeitet

+++ makesymlinks.sh
@@ -5,6 +5,7 @@
 ############################
 
 error="[\e[31merror\e[0m"
+info="[\e[32minfo\e[0m"
 move="[\e[32mmove\e[0m"
 delete="[\e[32mdelete\e[0m"
 create="[\e[32mcreate\e[0m"
@@ -20,13 +21,13 @@ folders=('gimp-2.8/scripts')
 
 # create dotfiles_old in homedir
 olddir=~/dotfiles_old             # old dotfiles backup directory
-  if [[ ! -L ~/.$file ]]
-  then
-    echo "$create $highlight$olddir$default] Creating for backup of any existing dotfiles in ~"
-  else
-    echo "$idle Folder $highlight$olddir$default for backup already exists"
-  fi
-  mkdir -p $olddir
+if [[ ! -L ~/.$file ]]
+then
+  echo "$create $highlight$olddir$default] Creating for backup of any existing dotfiles in ~"
+else
+  echo "$idle Folder $highlight$olddir$default for backup already exists"
+fi
+mkdir -p $olddir
 echo "...done\n"
 
 # change to the dotfiles directory
@@ -36,16 +37,21 @@ echo "...done\n"
 
 # create symlinks
 for file in $files; do
-  if [[ ! -L ~/.$file ]]
+  if [[ -e ~/.$file ]]
   then
-    echo "$move $highlight~/.$file$default] Move from ~ to $olddir"
-    mv ~/.$file $olddir/
+    if [[ ! -L ~/.$file ]]
+    then
+      echo "$move $highlight~/.$file$default] Move from ~ to $olddir"
+      mv ~/.$file $olddir/
+    else
+      echo "$delete $highlight~/.$file$default] is a symlink and will be deletet"
+      rm ~/.$file
+    fi
   else
-    echo "$delete $highlight~/.$file$default] is a symlink and will be deletet"
-    rm ~/.$file
+    echo "$info $highlight~/.$file$default] file doesnt exist"
   fi
-    echo "$create $highlight~/.$file$default] Creating symlink in home directory"
-    ln -s $dir/$file ~/.$file
+  echo "$create $highlight~/.$file$default] Creating symlink in home directory"
+  ln -s $dir/$file ~/.$file
 done
 
 # create symlinks
@@ -71,8 +77,8 @@ for folder in $folders; do
         echo "$error $highlight$folder$default] file already exists but its not a Folder!"
       fi
     else
-        echo "$create $highlight$folder$default] Creating Symlink in home directory"
-        ln -s -r $dir/$folder/ ~/.${folder%/*}/
+      echo "$create $highlight$folder$default] Creating Symlink in home directory"
+      ln -s -r $dir/$folder/ ~/.${folder%/*}/
     fi
   else
     echo "$idle symlink to $highlight$folder$default already exists"
